added version tag for building

+++ build.ps1
@@ -1,12 +1,15 @@
-$outputFile = "automatic-folder-reader.xpi"
 $buildDir = "build"
 $srcDir = "src"
+$appName = "automatic-folder-reader"
 
-if (Test-Path $buildDir) {
-    Remove-Item -Recurse -Force $buildDir
-}
+if (Test-Path $buildDir) { Remove-Item -Recurse -Force $buildDir }
 
 New-Item -ItemType Directory -Path $buildDir | Out-Null
+
+$manifest = Get-Content "$srcDir\manifest.json" | ConvertFrom-Json
+$version = $manifest.version
+
+$outputFile = "$appName-$version.xpi"
 
 Push-Location $srcDir
 
@@ -19,4 +22,3 @@ $items = @(
 )
 
 Compress-Archive -Path $items -DestinationPath "../$buildDir/$outputFile" -Force
-
